move library path to kong/dns
