+++ kong/resty/dns/client.lua
@@ -1,8 +1,8 @@
 -- Use the new dns client library instead. If you want to switch to the original
 -- one, you can set `legacy_dns_client = on` in kong.conf.
 if ngx.shared.kong_dns_cache and not _G.busted_legacy_dns_client then
-  package.loaded["kong.resty.dns_client"] = nil
-  return require("kong.resty.dns_client")
+  package.loaded["kong.dns.client"] = nil
+  return require("kong.dns.client")
 end
 
 --------------------------------------------------------------------------
