+++ spec/01-unit/30-new-dns-client/01-utils_spec.lua
@@ -1,4 +1,4 @@
-local utils = require "kong.resty.dns_client.utils"
+local utils = require "kong.dns.utils"
 local tempfilename = require("pl.path").tmpname
 local writefile = require("pl.utils").writefile
 local splitlines = require("pl.stringx").splitlines
