+++ spec/01-unit/30-new-dns-client/02-old_client_spec.lua
@@ -95,8 +95,8 @@ describe("[DNS client]", function()
     end
 
     -- restore its API overlapped by the compatible layer
-    package.loaded["kong.resty.dns_client"] = nil
-    client = require("kong.resty.dns_client")
+    package.loaded["kong.dns.client"] = nil
+    client = require("kong.dns.client")
     client.resolve = function (self, name, opts, tries)
       if opts and opts.return_random then
         return self:resolve_address(name, opts.port, opts.cache_only, tries)
